Refactor move InfoStatisticAndReview
