+++ src/components/user/info/info-statistic-and-review/InfoStatisticAndReview.tsx
