+++ src/pages/user/Info.tsx
@@ -2,7 +2,7 @@ import { colFlex } from '@styles/flexStyles';
 import AppContainer from '@components/common/container/AppContainer';
 import InfoMain from '@components/user/info/info-main/InfoMain';
 import InfoBenefit from '@components/user/info/info-benefit/InfoBenefit';
-import InfoStatisticAndReview from '@components/user/info/InfoStatisticAndReview';
+import InfoStatisticAndReview from '@components/user/info/info-statistic-and-review/InfoStatisticAndReview';
 import InfoOurService from '@components/user/info/InfoOurService';
 import InfoCreatorsWords from '@components/user/info/InfoCreatorsWords';
